move eeprom code from include/pipettes to include/eeprom
