+++ include/can/core/messages.hpp
@@ -6,8 +6,8 @@
 #include <span>
 
 #include "can/core/ids.hpp"
+#include "eeprom//core/types.hpp"
 #include "common/core/bit_utils.hpp"
-#include "common/core/eeprom.hpp"
 #include "common/core/version.h"
 #include "parse.hpp"
