+++ include/eeprom/core/types.hpp
